BACKEND/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pillow
